Refactor the result back into the library
