+++ src/py-opentimelineio/opentimelineio/__init__.py
@@ -43,4 +43,5 @@ from . import (
     adapters,
     hooks,
     algorithms,
+    url_utils
 )
