+++ tests/test_url_conversions.py
@@ -1,0 +1,77 @@
+#
+# Copyright Contributors to the OpenTimelineIO project
+#
+# Licensed under the Apache License, Version 2.0 (the "Apache License")
+# with the following modification; you may not use this file except in
+# compliance with the Apache License and the following modification to it:
+# Section 6. Trademarks. is deleted and replaced with:
+#
+# 6. Trademarks. This License does not grant permission to use the trade
+#    names, trademarks, service marks, or product names of the Licensor
+#    and its affiliates, except as required to comply with Section 4(c) of
+#    the License and to reproduce the content of the NOTICE file.
+#
+# You may obtain a copy of the Apache License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the Apache License with the above modification is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the Apache License for the specific
+# language governing permissions and limitations under the Apache License.
+#
+
+""" Unit tests of functions that convert between file paths and urls. """
+
+import unittest
+import os
+
+import opentimelineio as otio
+
+SAMPLE_DATA_DIR = os.path.join(os.path.dirname(__file__), "sample_data")
+SCREENING_EXAMPLE_PATH = os.path.join(SAMPLE_DATA_DIR, "screening_example.edl")
+MEDIA_EXAMPLE_PATH_REL = os.path.relpath(
+    os.path.join(
+        os.path.dirname(__file__),
+        "..",  # root
+        "docs",
+        "_static",
+        "OpenTimelineIO@3xDark.png"
+    )
+)
+MEDIA_EXAMPLE_PATH_URL_REL = otio.url_utils.url_from_filepath(
+    MEDIA_EXAMPLE_PATH_REL
+)
+MEDIA_EXAMPLE_PATH_ABS = os.path.abspath(
+    MEDIA_EXAMPLE_PATH_REL.replace(
+        "3xDark",
+        "3xLight"
+    )
+)
+MEDIA_EXAMPLE_PATH_URL_ABS = otio.url_utils.url_from_filepath(
+    MEDIA_EXAMPLE_PATH_ABS
+)
+
+
+class TestConversions(unittest.TestCase):
+    def test_roundtrip_abs(self):
+        self.assertTrue(MEDIA_EXAMPLE_PATH_URL_ABS.startswith("file://"))
+        full_path = os.path.abspath(
+            otio.url_utils.filepath_from_url(MEDIA_EXAMPLE_PATH_URL_ABS)
+        )
+
+        # should have reconstructed it by this point
+        self.assertEqual(full_path, MEDIA_EXAMPLE_PATH_ABS)
+
+    def test_roundtrip_rel(self):
+        self.assertFalse(MEDIA_EXAMPLE_PATH_URL_REL.startswith("file://"))
+
+        result = otio.url_utils.filepath_from_url(MEDIA_EXAMPLE_PATH_URL_REL)
+
+        # should have reconstructed it by this point
+        self.assertEqual(result, MEDIA_EXAMPLE_PATH_REL)
+
+
+if __name__ == "__main__":
+    unittest.main()
